+++ src/app/children/attendance/add-day-attendance/add-day-attendance.component.ts
@@ -1,10 +1,10 @@
 import { Component, OnInit } from '@angular/core';
-import {ChildrenService, ChildWithRelation} from '../../children.service';
-import {Child} from '../../child';
+import {ChildrenService} from '../../children.service';
 import {FilterSelection} from '../../../ui-helper/filter-selection/filter-selection';
 import {EntityMapperService} from '../../../entity/entity-mapper.service';
 import {AttendanceDay, AttendanceStatus} from '../attendance-day';
 import {AttendanceMonth} from '../attendance-month';
+import {ChildWithRelation} from '../../childWithRelation';
 
 @Component({
   selector: 'app-add-day-attendance',
