+++ src/app/children/children.module.ts
@@ -39,7 +39,7 @@ import {
   MatIconModule, MatTooltipModule
 } from '@angular/material';
 import {FlexLayoutModule} from '@angular/flex-layout';
-import {ChildrenService, ChildWithRelation} from './children.service';
+import {ChildrenService} from './children.service';
 import {AttendanceBlockComponent} from './attendance/attendance-block/attendance-block.component';
 import {ChildAttendanceComponent} from './attendance/child-attendance/child-attendance.component';
 import {UiHelperModule} from '../ui-helper/ui-helper.module';
